refactor: moved locationUtils and changed name of googleMaps file
moved locationUtils to correct folder and changed googleMaps file to follow our conventions
