+++ src/lib/locationUtils.ts
